# docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mini Ticketing API",
        "description": "Event ticketing API built with Laravel 13, MongoDB and Passport OAuth2.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8004/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "List events",
                "description": "Public endpoint. Optionally filter by city.",
                "operationId": "3befe2e8582683a8dc70f7b706a5a97d",
                "parameters": [
                    {
                        "name": "city",
                        "in": "query",
                        "description": "Filter events by city",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Riyadh"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A paginated list of events"
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Create an event",
                "description": "Requires a token carrying the events:create scope. A regular user's token receives 403.",
                "operationId": "1ed79d9ab6d41b87387cad6237213e5d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "city",
                                    "date"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Riyadh Season Concert"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Riyadh"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-11-20"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "active",
                                        "enum": [
                                            "active",
                                            "paused",
                                            "cancelled"
                                        ]
                                    },
                                    "meta": {
                                        "description": "Free-form data that varies by event type",
                                        "type": "object",
                                        "example": {
                                            "artist": "Mohammed Abdu",
                                            "venue": "Mrsool Park"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Event created"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Token lacks the events:create scope"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Get one event",
                "description": "Public. Includes the event's ticket types.",
                "operationId": "abf4f4bbfe75221b5bab720a701f17b8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The event id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "6a842696e4e5ed74df0bee52"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The event with its tickets"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete an event",
                "description": "Only possible when the event has no orders. Cancel it instead — deleting would orphan those orders and erase a financial record.",
                "operationId": "54be98df9381a8b89013d29e170a31fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted, along with its ticket types"
                    },
                    "403": {
                        "description": "Token lacks the events:manage scope"
                    },
                    "422": {
                        "description": "The event has orders"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Events"
                ],
                "summary": "Update an event",
                "description": "Partial update — only the fields you send are changed. Requires the events:manage scope.",
                "operationId": "b374fc5e8a4d9d816f3a4de49c6c17ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Riyadh Season Concert"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Riyadh"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-11-20"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "paused",
                                            "cancelled"
                                        ]
                                    },
                                    "meta": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Event updated"
                    },
                    "403": {
                        "description": "Token lacks the events:manage scope"
                    },
                    "404": {
                        "description": "Event not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Obtain a token from /login, then send it as: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}